Fix click to copy course code not working in course calendar event popup
