+++ client/src/components/Calendar/CourseCalendarEvent.js
@@ -8,6 +8,7 @@ import locations from '../SectionTable/static/locations.json';
 import { deleteCourse, deleteCustomEvent } from '../../actions/AppStoreActions';
 import CustomEventDialog from '../CustomEvents/CustomEventDialog';
 import AppStore from '../../stores/AppStore';
+import { clickToCopy } from '../../helpers';
 
 const styles = {
     courseContainer: {
@@ -60,17 +61,6 @@ const styles = {
             borderRadius: '50%',
         },
     },
-};
-
-const clickToCopy = (event, code) => {
-    event.stopPropagation();
-
-    let inputElem = document.createElement('input');
-    document.body.appendChild(inputElem);
-    inputElem.setAttribute('value', code);
-    inputElem.select();
-    document.execCommand('copy');
-    document.body.removeChild(inputElem);
 };
 
 const genMapLink = (location) => {
