+++ client/src/helpers.js
@@ -1,3 +1,5 @@
+import { openSnackbar } from './actions/AppStoreActions';
+
 export async function getCoursesData(userData) {
     const dataToSend = {};
     const addedCourses = [];
@@ -63,4 +65,16 @@ export async function getCoursesData(userData) {
         addedCourses: addedCourses,
         customEvents: userData.customEvents,
     };
+}
+
+export function clickToCopy(event, sectionCode) {
+    event.stopPropagation();
+
+    let tempEventTarget = document.createElement('input');
+    document.body.appendChild(tempEventTarget);
+    tempEventTarget.setAttribute('value', sectionCode);
+    tempEventTarget.select();
+    document.execCommand('copy');
+    document.body.removeChild(tempEventTarget);
+    openSnackbar('success', 'Section code copied to clipboard');
 }
